refactor(A-Packages): Organise duke classes to packages
Closes #5
