+++ src/main/java/duke/command/ByeCommand.java
@@ -1,3 +1,8 @@
+package duke.command;
+
+import duke.task.TaskList;
+import duke.ui.Ui;
+
 public class ByeCommand extends Command {
   public ByeCommand(String commandString) {
     super(CommandType.BYE, commandString);
