+++ src/main/java/duke/command/CommandType.java
@@ -1,3 +1,5 @@
+package duke.command;
+
 public enum CommandType {
   LIST_DATE("date"),
   LIST("list"),
