+++ src/main/java/duke/command/DeleteCommand.java
@@ -1,3 +1,10 @@
+package duke.command;
+
+import duke.exception.DukeException;
+import duke.task.Task;
+import duke.task.TaskList;
+import duke.ui.Ui;
+
 public class DeleteCommand extends UpdateCommand {
   public DeleteCommand(String commandString) {
     super(CommandType.DELETE, commandString);
