+++ src/main/java/duke/command/DoneCommand.java
@@ -1,3 +1,10 @@
+package duke.command;
+
+import duke.exception.DukeException;
+import duke.task.Task;
+import duke.task.TaskList;
+import duke.ui.Ui;
+
 public class DoneCommand extends UpdateCommand {
   public DoneCommand(String commandString) {
     super(CommandType.DONE, commandString);
