+++ src/main/java/duke/command/UpdateCommand.java
@@ -1,3 +1,8 @@
+package duke.command;
+
+import duke.exception.DukeException;
+import duke.parser.Parser;
+
 public abstract class UpdateCommand extends Command {
   public UpdateCommand(CommandType commandType, String commandString) {
     super(commandType, commandString);
