+++ src/main/java/duke/exception/DukeException.java
@@ -1,3 +1,5 @@
+package duke.exception;
+
 public class DukeException extends Exception {
   private static final long serialVersionUID = 50572093442410692L;
