+++ src/main/java/duke/parser/Parser.java
@@ -1,3 +1,9 @@
+package duke.parser;
+
+import duke.command.Command;
+import duke.command.CommandType;
+import duke.exception.DukeException;
+import duke.ui.Ui;
 import java.time.LocalDate;
 import java.time.format.DateTimeParseException;
 import java.util.Optional;
